rename : PointUtil 디렉토리 변경으로 인한 변경
# Conflicts:
#	src/test/java/org/programmers/signalbuddyfinal/domain/postit/service/PostItServiceTest.java
